requirements.txt
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.4
python-dotenv==1.2.2
python-telegram-bot==21.11.1
requests==2.33.1
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
